Moved logic into own library (modular)
