+++ .gitignore
@@ -1,5 +1,5 @@
 .vscode/
-build
+build/
 *Thumbs.db
 *.swp
 *.aps
@@ -7,3 +7,5 @@ res/settings.ini
 res/KLS_Logic.chw
 /.vs
 /CMakeSettings.json
+logic/build/
+gui/build/
