+++ include/gui/MainApp.h
@@ -20,7 +20,7 @@
 #include "wx/fs_zip.h"
 #include "threadLogic.h"
 #include "autoSaveThread.h"
-#include "../logic/logic_values.h"
+#include "logic_values.h"
 #include "LibraryParse.h"
 #include "gl_defs.h"
 #include "klsMessage.h"
