+++ include/gui/OscopeCanvas.h
@@ -17,7 +17,7 @@ class OscopeCanvas;
 #include "MainApp.h"
 #include "wx/glcanvas.h"
 #include "GUICircuit.h"
-#include "../logic/logic_values.h"
+#include "logic_values.h"
 
 #include <map>
 #include <vector>
