+++ include/gui/wireSegment.h
@@ -9,7 +9,7 @@
 #include <cfloat>
 #include <string>
 #include <vector>
-#include "../logic/logic_values.h" // StateType
+#include "logic_values.h" // StateType
 #include "klsCollisionChecker.h"
 
 class guiGate;
