+++ logic/include/logic_circuit.h
