+++ logic/include/logic_gate.h
