+++ logic/include/logic_junction.h
