+++ logic/include/logic_wire.h
