+++ logic/src/logic_circuit.cpp
