+++ logic/src/logic_gate.cpp
