+++ logic/src/logic_junction.cpp
