+++ logic/src/logic_wire.cpp
